pyproject.toml
[tool.pdm.dev-dependencies]
dev = [
    "black<25.0.0,>=24.2.0",
    "ruff<1.0.0,>=0.3.0",
    "mypy<2.0.0,>=1.8.0",
    "pre-commit<4.0.0,>=3.6.0",
    "types-requests<3.0.0.0,>=2.32.0.20241016",
]

[tool.pdm.build]
includes = []
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.logfire]


[project]
authors = [
    {name = "Jun Liu", email = "antyser@gmail.com"},
]
requires-python = "<4.0,>=3.12"
dependencies = [
    "agentql<2.0.0,>=1.6.2",
    "anthropic>=0.43.1",
    "beautifulsoup4>=4.12.3",
    "brotli>=1.1.0",
    "chardet>=5.2.0",
    "datasets>=3.2.0",
    "duckduckgo-search>=7.3.0",
    "fake-headers>=1.0.2",
    "fastapi>=0.115.7",
    "google-generativeai>=0.8.4",
    "html5lib>=1.1",
    "httpx[http2]==0.27.2",
    "instructor>=1.7.2",
    "ipython>=8.31.0",
    "jupyter>=1.1.1",
    "logfire[fastapi,httpx,system-metrics]>=3.1.0",
    "loguru<1.0.0,>=0.7.3",
    "markdown>=3.7",
    "markdownify>=0.14.1",
    "openai>=1.59.8",
    "openai-agents>=0.0.4",
    "playwright>=1.49.1",
    "playwright-dompath>=0.0.1",
    "pydantic-ai[logfire]>=0.0.19",
    "pydantic-settings>=2.7.1",
    "python-dotenv<2.0.0,>=1.0.1",
    "requests>=2.32.3",
    "sentry-sdk>=2.20.0",
    "smolagents>=1.10.0",
    "streamlit>=1.41.1",
    "supabase>=2.11.0",
    "tiktoken>=0.8.0",
    "types-requests>=2.32.0.20241016",
    "unstructured>=0.16.12",
    "uvicorn>=0.34.0",
    "youtube-transcript-api>=0.6.3",
]
name = "shop-backend"
version = "0.1.0"
description = ""
readme = "README.md"

[dependency-groups]
dev = [
    "black<25.0.0,>=24.2.0",
    "ruff<1.0.0,>=0.3.0",
    "mypy<2.0.0,>=1.8.0",
    "pre-commit<4.0.0,>=3.6.0",
    "types-requests<3.0.0.0,>=2.32.0.20241016",
]

[tool.black]
line-length = 150
target-version = ["py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 150
target-version = "py312"
lint.select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
lint.ignore = []

# Exclude files in tests/ directory from ruff checking
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "app/tests/**",
]

